+++ free-ton/contracts/staking/interfaces/IElection.sol
@@ -5,14 +5,14 @@ interface IElection {
     struct MembershipRequest {
         address staker_addr;
         uint256 ton_pubkey;
-        uint256 eth_addr;
+        uint160 eth_addr;
         uint128 tokens;
     }
 
     function applyForMembership(
         address staker_addr,
         uint256 ton_pubkey,
-        uint256 eth_address,
+        uint160 eth_address,
         uint128 tokens,
         uint128 lock_time,
         address send_gas_to,
